services/banana/banana_requirements.txt
git+https://github.com/samhita-alla/katna.git#egg=katna
pandas
pytorch-lightning
s2sphere
scikit-learn
torchvision
msgpack
plotly
geopy
opentelemetry-sdk==1.12.0
opentelemetry-semantic-conventions==0.33b0
Pillow
sanic==22.6.2
banana-dev
matplotlib
onnxruntime-gpu==1.12
boto3
smart_open
python-dotenv
